Extract color UI to same class.
